Performance optimization and various things
- resizing icons takes too much time to let it loop, we do it once and copy screen to tex, selection overlays are then drawn in a loop
- fixed user menu list to match title menu
- fixed stuff with color swapping and enabled to swap alpha pixels too

regression : touch not working on user / title menus ^^, will need to check why

+++ inc/menu.h
@@ -55,7 +55,7 @@ namespace ui
             //Selected + frame counting for auto-scroll
             int selected = 0, fc = 0, start = 0, retEvent = MENU_NOTHING, fontSize = 19;
             //How much we shift the color of the rectangle
-            uint8_t clrSh = 0;
+            int clrSh = 0;
             bool clrAdd = true;
             bool separate = true;
 			bool leftmenu = false;
